pyproject.toml
[project]
name = "oceans-of-nyc"
version = "0.1.0"
description = "A community scavenger hunt to spot all the Fisker Oceans operating in NYC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "atproto>=0.0.55",
    "python-dotenv>=1.0.0",
    "staticmap>=0.5.7",
    "psycopg2-binary>=2.9.10",
    "modal>=1.2.4",
    "twilio>=9.0.0",
    "imagehash>=4.3.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
]

[project.scripts]
ocean-bot = "main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["post", "chat", "data", "process", "database", "validate", "geolocate", "notify"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# Enable recommended rule sets
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modernize syntax)
    "B",   # flake8-bugbear (common bugs)
    "A",   # flake8-builtins (shadowing builtins)
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "ARG", # flake8-unused-arguments
]

# Ignore specific rules if needed (can enable these incrementally)
ignore = [
    "E501",  # line too long (formatter handles this)
    "E731",  # lambda assignment (sometimes clearer than def)
    "ARG001",  # unused function argument (common in callbacks/handlers)
    "ARG002",  # unused method argument
    "B904",  # raise from within except (can add later)
    "SIM102",  # nested if simplification (readability preference)
    "SIM108",  # ternary operator suggestion (readability preference)
    "SIM117",  # nested with statements (readability preference)
    "RET504",  # unnecessary assignment before return (sometimes clearer)
    "C401",  # generator to comprehension (minor optimization)
    "C414",  # unnecessary list() call (minor optimization)
    # "F841",  # unused local variable (can clean up later)
]

[tool.ruff.lint.per-file-ignores]
# Allow unused imports in __init__.py files
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true

# Gradual adoption - disable strict checks initially
# Re-enable these incrementally as types are added
no_implicit_optional = false  # Allow implicit Optional (e.g., arg: str = None)
warn_return_any = false  # Allow returning Any from typed functions
warn_unused_ignores = false  # Don't warn about unused type ignores
strict_equality = false  # Allow loose equality comparisons

# Allow untyped code for now
disallow_untyped_defs = false
disallow_any_generics = false
check_untyped_defs = false

# Ignore missing imports for libraries without stubs
ignore_missing_imports = true

# Per-module options to ignore specific issues
[[tool.mypy.overrides]]
module = [
    "post.bluesky",
    "geolocate.exif",
    "geolocate.maps",
    "geolocate.geocoding",
    "chat.session",
    "chat.webhook",
    "database.models",
    "validate.tlc",
    "main",
    "modal_app",
]
# Disable specific checks for these modules with complex type issues
# Can re-enable these incrementally as types are fixed
disable_error_code = [
    "attr-defined",     # Missing attributes on objects
    "no-redef",         # Duplicate definitions
    "assignment",       # Type mismatches in assignments
    "arg-type",         # Argument type mismatches
    "return-value",     # Return type mismatches
    "import-untyped",   # Untyped library imports
]
